kiddyer-server/requirements.txt
django
django-rest-framework
mysqlclient